changed a few datatypes and created controllers for corresponding entities

+++ backend/spring-boot-salonmanagementapp/src/main/java/com/codigoslatinos/springbootsalonmanagementapp/dao/ZipCodeRepository.java
@@ -3,6 +3,6 @@ package com.codigoslatinos.springbootsalonmanagementapp.dao;
 import com.codigoslatinos.springbootsalonmanagementapp.entity.ZipCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ZipCodeRepository extends JpaRepository<ZipCode, Integer> {
-    void deleteByZipcode(int zipcode);
+public interface ZipCodeRepository extends JpaRepository<ZipCode, Long> {
+    void deleteByZipcode(Long zipcode);
 }
